README: Fix the code example to use the v2 API.
For now, the code example in README.md does not work because it
relies on the old API (which has been deprecated since v2.0.0).

This small patch fixes this issue by migrating the code to use the
newer interfaces. After this commit, the code example will work
fine as follows:

    [log-test] $ go build
    [log-test] $ ./log-test
    [log-test] $ cat logs/access_log.201712280900
    2017/12/28 15:37:48 Hello, World!